+++ lib/components/contact_option.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/src/foundation/key.dart';
+import 'package:flutter/src/widgets/container.dart';
+import 'package:flutter/src/widgets/framework.dart';
+
+// Placeholder class, as we haven't planned on how to fetch contact
+class ContactOption extends StatelessWidget {
+  const ContactOption({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
